Feature : Allow resizing of files before downloading (#22)
* chore: Update `bundler`, `rake` and `Gemfile.lock`

Updates the gem dependencies.
Update bundler and rake.

* feat: Add resize_to_limit filter

Allow resizing of zipped files before downloading.

* fix: Fix dependency on `service_url`

According to the documentation, `service_url` shouldn't be used
directly. Replace the call to `service_url` with `url_for`.

* fix: send `url_for` isn't supported by all services

Sending the `url_for` method isn't supported by all services.

* chore: Update `CHANGELOG.yml`

Update the `CHANGELOG.yml` for version 0.3.4 with the resize feature.

* chore: Update `README.md`

Update the `README.md` with details on how to use the resize feature.

* chore: Bump version number

Increase the version number to 0.3.4.
Add my name to the authors.

* chore: Document `resize_to_limit` filter

Document the `resize_to_limit` filter that can be applied to images
before being zipped.

* feat: Add a test for `resize_to_limit`

Add a test to ensure the behavior of resize_to_limit is working.